improvement(product-offer): text+kafka publishers, corda class whitelisting, corda jar sealing fixes (#7)
* improvement(product-offer): text+kafka publishers, corda class whitelisting, corda jar sealing fixes

Added Kafka and Text product offer update listeners to the spring api. Fixed class whitelistingas
previous use of reflection to generate whitelist for nextwork info model library is not allowed on
corda dlt nodes due to deterministic JVM. Also exclude spring data dependencies from the nextworks
library as if included the spring dependencies are included in the corda contracts cordapp jar which
is sealed. Meaning when jar is loaded in the spring api and the spring dependencies are loaded
before the cordapp contract jar then security exception is thrown as the package names have been
sealed by the contracts cordapp jar. Excluding causes a few compile warnings however since the jpa
uses are limited to the annotations on info model classes this doesn't cause an issue currently. *
Need to determine a more long term fix to this issue.

* build(fix): added java-library plugin so dlt-api module jar is compiled correctly

* fix(tests): fixing test config and corda module load on startup